fix(inventario): la cantidad de movimientos pasa a decimal
El escritorio migro las cantidades de entero a numeric(14,3) para poder
facturar metros de tejido, kilos y litros. La app seguia esperando un
int en InventoryMovement.quantity, y Prisma serializa un Decimal como
CADENA en JSON: la pantalla de movimientos de un articulo habria
reventado al parsear "1.000" como entero.

InventoryMovement.quantity pasa a double con el conversor toDouble, que
ya aguanta num y String. createInventoryEntry acepta double.

El formulario de entrada de stock validaba con int.tryParse, o sea que
teclear 1,5 daba "Introduce cantidad valida". Ahora acepta decimales, con
coma o con punto, y el teclado abre en modo decimal.

flutter analyze: sin incidencias.

+++ lib/core/api/endpoints.dart
@@ -675,7 +675,7 @@ class Endpoints {
 
   Future<void> createInventoryEntry({
     required int productId,
-    required int quantity,
+    required double quantity,
     required double unitCost,
     String? notes,
   }) async {
